Creazione script importazione certificati

+++ app/Models/Certificate.php
@@ -56,7 +56,9 @@ class Certificate extends Model
         });
 
         static::forceDeleted(function($model){
-            Storage::delete($model->document);
+            if($model->document){
+                Storage::delete($model->document);
+            }
         });
     }
